Jobs Code Refactor
