+++ lib/frontend/screens/functionality/job/apply/appliedCards.dart
