+++ lib/frontend/screens/functionality/job/apply/appliedjobdetail.dart
